Fix challenge tags displaying incorrectly.
Challenge tag IDs are specified using the "_id" field in the REST API, but the Swift code looks
for an "id" field since the Swift field is named "id". This causes every tag to use the nil ID,
which means that the most recently seen tag overwrites all previous tags. This commit fixes
that by explicitly using the "_id" field from the JSON representation.